medical_triage_agent/requirements.txt
google-adk>=0.3.0
google-genai>=1.0.0
pydantic>=2.0
numpy>=1.24
# Optional but recommended for better retrieval quality:
# sentence-transformers>=2.2
